pyproject.toml
[project]
name = "fast-api-uv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "fastapi>=0.116.1",
    "passlib>=1.7.4",
    "phonenumbers>=9.0.13",
    "pydantic-extra-types>=2.10.5",
    "pydantic-settings>=2.10.1",
    "pydantic[email]>=2.11.7",
    "python-dateutil>=2.9.0.post0",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]
